+++ include/x308/SystemStatusPresenter.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "x308/SystemStatusReport.hpp"
+
+#include <iosfwd>
+
+namespace x308 {
+
+class SystemStatusPresenter {
+public:
+    static void print(const SystemStatusReport& report, std::ostream& output);
+};
+
+}  // namespace x308